work around GitHub Action security guarantees
GitHub Actions aren't allowed to change any of the integrations in
.github/workflows. So, we have to work around that by creating a new "versions"
file that will be explicitly included in to be used in `actions/setup-go`.

The current file is expected to be at .github/versions/go and the format is a
single plaintext version to included. That's in order to support it being
trivially `cat`'ed in the GitHub Action that is also using `actions/setup-go`.

I made https://github.com/actions/setup-go/issues/23 in hopes that I can give
`actions/setup-go` a PR to make this a baked-in concept.